4.21 created

+++ Osa4/blogilista-backend/controllers/blogs.js
@@ -52,14 +52,29 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 });
 
+
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({
+        error: 'token missing or invalid'
+      });
+    }
+
     const blog = await Blog.findById(request.params.id);
+
     if (!blog) {
       return response.status(404).json({ error: 'Blog not found' });
     }
-    await blog.deleteOne();
-    response.status(204).end();
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await blog.deleteOne();
+      response.status(204).end();
+    } else {
+      response.status(401).json({ error: 'token invalid' });
+    }
   } catch (error) {
     next(error);
   }
